+++ instrumentation/opentelemetry-instrumentation-aws-lambda/tests/test_aws_lambda_instrumentation_manual.py
@@ -587,8 +587,6 @@ class TestAwsLambdaInstrumentor(TestAwsLambdaInstrumentorBase):
         "os.environ",
         {
             OTEL_INSTRUMENTATION_AWS_LAMBDA_FORCE_FLUSH: "false",
-            OTEL_INSTRUMENTATION_AWS_LAMBDA_FLUSH_METRICS: "false",
-            OTEL_INSTRUMENTATION_AWS_LAMBDA_FLUSH_TRACES: "false",
         },
     )
     def test_force_flush_disabled_globally(
@@ -612,7 +610,6 @@ class TestAwsLambdaInstrumentor(TestAwsLambdaInstrumentorBase):
         {
             OTEL_INSTRUMENTATION_AWS_LAMBDA_FORCE_FLUSH: "false",
             OTEL_INSTRUMENTATION_AWS_LAMBDA_FLUSH_TRACES: "true",
-            OTEL_INSTRUMENTATION_AWS_LAMBDA_FLUSH_METRICS: "false",
         },
     )
     def test_force_flush_selective_traces(
@@ -635,7 +632,6 @@ class TestAwsLambdaInstrumentor(TestAwsLambdaInstrumentorBase):
         "os.environ",
         {
             OTEL_INSTRUMENTATION_AWS_LAMBDA_FORCE_FLUSH: "false",
-            OTEL_INSTRUMENTATION_AWS_LAMBDA_FLUSH_TRACES: "false",
             OTEL_INSTRUMENTATION_AWS_LAMBDA_FLUSH_METRICS: "true",
         },
     )
@@ -656,6 +652,50 @@ class TestAwsLambdaInstrumentor(TestAwsLambdaInstrumentorBase):
     @mock.patch("opentelemetry.instrumentation.aws_lambda.get_meter_provider")
     @mock.patch("opentelemetry.instrumentation.aws_lambda.get_tracer_provider")
     @mock.patch.dict(
+        "os.environ",
+        {
+            OTEL_INSTRUMENTATION_AWS_LAMBDA_FLUSH_TRACES: "false",
+        },
+    )
+    def test_force_flush_disabled_selective_traces(
+        self, mock_get_tracer_provider, mock_get_meter_provider
+    ):
+        mock_tracer_provider = mock.Mock()
+        mock_meter_provider = mock.Mock()
+        mock_get_tracer_provider.return_value = mock_tracer_provider
+        mock_get_meter_provider.return_value = mock_meter_provider
+
+        AwsLambdaInstrumentor().instrument()
+        mock_execute_lambda()
+
+        mock_tracer_provider.force_flush.assert_not_called()
+        mock_meter_provider.force_flush.assert_called_once()
+
+    @mock.patch("opentelemetry.instrumentation.aws_lambda.get_meter_provider")
+    @mock.patch("opentelemetry.instrumentation.aws_lambda.get_tracer_provider")
+    @mock.patch.dict(
+        "os.environ",
+        {
+            OTEL_INSTRUMENTATION_AWS_LAMBDA_FLUSH_METRICS: "false",
+        },
+    )
+    def test_force_flush_disabled_selective_metrics(
+        self, mock_get_tracer_provider, mock_get_meter_provider
+    ):
+        mock_tracer_provider = mock.Mock()
+        mock_meter_provider = mock.Mock()
+        mock_get_tracer_provider.return_value = mock_tracer_provider
+        mock_get_meter_provider.return_value = mock_meter_provider
+
+        AwsLambdaInstrumentor().instrument()
+        mock_execute_lambda()
+
+        mock_tracer_provider.force_flush.assert_called_once()
+        mock_meter_provider.force_flush.assert_not_called()
+
+    @mock.patch("opentelemetry.instrumentation.aws_lambda.get_meter_provider")
+    @mock.patch("opentelemetry.instrumentation.aws_lambda.get_tracer_provider")
+    @mock.patch.dict(
         "os.environ", {OTEL_INSTRUMENTATION_AWS_LAMBDA_FORCE_FLUSH: "true"}
     )
     def test_force_flush_instrumentation_options_override(
@@ -670,7 +710,6 @@ class TestAwsLambdaInstrumentor(TestAwsLambdaInstrumentorBase):
         AwsLambdaInstrumentor().instrument(
             force_flush=False,
             flush_metrics=True,
-            flush_traces=False,
         )
         mock_execute_lambda()